pyproject.toml
[tool.poetry]
name = "genai-test"
version = "0.1.0"
description = "Test code for various genai experiments"
authors = ["Rajesh Korde <rajesh.korde@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
lancedb = "^0.13.0"
sentence-transformers = "^3.1.1"
wikipedia-api = "^0.7.1"
tantivy = "^0.22.0"
duckdb = "^1.1.1"
pandas = "^2.2.3"
openai = "^1.50.1"
python-dotenv = "^1.0.1"
jupyter = "^1.1.1"
ipywidgets = "^8.1.5"
widgetsnbextension = "^4.0.13"
docx2txt = "^0.8"
datasets = "^3.0.1"
nltk = "^3.9.1"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
